requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Django==3.1.7
django-autoslug==1.9.8
django-widget-tweaks==1.4.8
flake8-django==1.1.1
gunicorn==20.0.4
iniconfig==1.1.1
isort==5.7.0
mccabe==0.6.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
